refactor(package): rename and move UpcApi
